refactor(import): remove unused error code and update related tests
- eliminate fileReadFailed error code from ImportDocumentErrorCode
- update tests to reflect the change in error handling
- restructure GoogleAuthError imports for consistency
