+++ src/features/importDocument/importDocumentErrorCode.ts
@@ -3,6 +3,5 @@
  */
 export enum ImportDocumentErrorCode {
   fileOpenFailed = 'importDocument.fileOpenFailed',
-  fileReadFailed = 'importDocument.fileReadFailed',
   documentImportFailed = 'importDocument.documentImportFailed',
 }
