+++ src/shared/lib/googleAuth/index.ts
@@ -1,0 +1,1 @@
+export { GoogleAuthError, GoogleAuthErrorCode } from './googleAuthError';
