+++ src/shared/service/repositories/index.ts
@@ -7,4 +7,5 @@ export {
   isRepositoryMarkerFileName,
   shouldHideRepositoryStorageFile,
 } from './repositoryStorageFiles';
-export type { RepositoryDirectoryEntry, RepositoryFacts } from './repositoryStorageFiles';
+export type { RepositoryDirectoryEntry } from './repositoryContracts';
+export type { RepositoryFacts } from './repositoryStorageFiles';
