+++ src/shared/service/repositories/repositoryContracts.ts
@@ -1,2 +1,6 @@
+import type { FSNodeStat } from '@shared/lib/virtualFileSystem';
+
 export { RepositoryImportErrorCode } from './repositoryImportErrorCode';
-export type { RepositoryDirectoryEntry } from './repositoryStorageFiles';
+
+/** Directory entry contract used for repository storage inspection and file-list filtering. */
+export type RepositoryDirectoryEntry = readonly [name: string, stat: FSNodeStat];
